refactor(dep-005b): migrate verification imports from schemas.workflow to schemas.verification
Migrate all callers of schemas.workflow.verification and schemas.workflow
verification-related re-exports to canonical schemas.verification imports.
VerificationResultSet migrated to schemas.results (its canonical home).

40+ files across production src, tests, notebooks, and docs. Bridge files
in schemas/workflow/verification/ preserved for dep-005c cleanup.
