+++ src/karenina/benchmark/__init__.py
@@ -1,7 +1,7 @@
 """Benchmark module for Karenina verification system."""
 
 from ..schemas.config import ModelConfig
-from ..schemas.workflow import FinishedTemplate, VerificationConfig, VerificationJob, VerificationResult
+from ..schemas.verification import FinishedTemplate, VerificationConfig, VerificationJob, VerificationResult
 from .benchmark import Benchmark
 from .verification import export_verification_results_csv, export_verification_results_json
 from .verification.runner import run_single_model_verification as run_question_verification
