+++ src/karenina/benchmark/verification/executor.py
@@ -25,8 +25,8 @@ from anyio.from_thread import start_blocking_portal
 if TYPE_CHECKING:
     from anyio.from_thread import BlockingPortal
 
+from ...schemas.verification import VerificationResult
 from ...schemas.verification.config import DEFAULT_ASYNC_MAX_WORKERS
-from ...schemas.workflow import VerificationResult
 from ...utils.answer_cache import AnswerTraceCache
 
 logger = logging.getLogger(__name__)
