+++ src/karenina/benchmark/verification/stages/core/orchestrator.py
@@ -7,7 +7,7 @@ import logging
 import time
 
 from .....schemas.domain import Rubric
-from .....schemas.workflow import VerificationResult
+from .....schemas.verification import VerificationResult
 from ..pipeline.abstention_check import AbstentionCheckStage
 from ..pipeline.deep_judgment_autofail import DeepJudgmentAutoFailStage
 from ..pipeline.deep_judgment_rubric_auto_fail import DeepJudgmentRubricAutoFailStage
