+++ tests/unit/test_rubric_results_deep_judgment.py
@@ -11,16 +11,17 @@ import json
 
 import pytest
 
-# Import and rebuild RubricJudgmentResults to resolve forward references
-from karenina.schemas.verification.model_identity import ModelIdentity
-from karenina.schemas.workflow.rubric_judgment_results import RubricJudgmentResults
-from karenina.schemas.workflow.rubric_results import RubricResults
-from karenina.schemas.workflow.verification import (
+from karenina.schemas.verification import (
     VerificationResult,
     VerificationResultDeepJudgmentRubric,
     VerificationResultMetadata,
     VerificationResultRubric,
 )
+
+# Import and rebuild RubricJudgmentResults to resolve forward references
+from karenina.schemas.verification.model_identity import ModelIdentity
+from karenina.schemas.workflow.rubric_judgment_results import RubricJudgmentResults
+from karenina.schemas.workflow.rubric_results import RubricResults
 
 RubricJudgmentResults.model_rebuild()
